requirements.txt
google-genai>=1.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
ddtrace>=2.0.0
braintrust>=0.0.100
sentence-transformers>=2.2.0
pydantic>=2.0.0
numpy>=1.24.0
